raspberry/bridge_slave/requirements.txt
pyserial
bleak
pyserial-asyncio
paho-mqtt
python-dotenv